WIP: removed webpack-bundles file. Consolidated the temporary-package dir
