+++ test/screenshot/App.js
@@ -2,9 +2,7 @@ import React from 'react';
 import {BrowserRouter, Route, Switch, withRouter} from 'react-router-dom';
 
 import Catalog from './Catalog';
-import TempPackage from './temporary-package';
-import FooTest from './temporary-package/FooTest';
-import BarTest from './temporary-package/BarTest';
+import TemporaryPackage from './temporary-package';
 
 export default class App extends React.Component {
   render() {
@@ -23,9 +21,7 @@ class RouteContainer_ extends React.Component {
         <nav className="mdc-toolbar-fixed-adjust">
           <Switch>
             <Route exact path="/" component={Catalog} />
-            <Route exact path="/temporary-package" component={TempPackage} />
-            <Route path="/temporary-package/foo-test" component={FooTest} />
-            <Route path="/temporary-package/bar-test" component={BarTest} />
+            <Route path="/temporary-package" component={TemporaryPackage} />
           </Switch>
         </nav>
       </main>
